+++ drizzle/0010_bumpy_shard.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "publishingJobs" ADD COLUMN "asset_url" text;